Updated carry or not

+++ C++ Programs/Carry or Not.cpp
@@ -3,7 +3,7 @@ using namespace std;
 int main()
 {
 	double a,b;
-	int c;
+	double c;
 	cin>>a>>b;
 	c=pow(a,b);
 	cout<<c;
